consertar bug cliente_x.txt

+++ ATV6/src/Estabelecimento.cpp
@@ -17,7 +17,7 @@ Estabelecimento::~Estabelecimento()
 }
 
 int Estabelecimento::venda(int codigo){
-  for(int i = 0; i < produtos.getSize(); i++){
+  for(size_t i = 0; i < produtos.getSize(); i++){
     if(codigo == produtos.at(i).codigo){
       if(produtos.at(i).quantidade == 0){
         std::cout << "Estoque esgotado :(" << std::endl;
